requirements.txt
Flask
flask-cors
google-generativeai
pypdf
python-docx
